I have identified an error in the collision kernel. I am not using atom ids for selecting velocities :-(

+++ include/cuda_dsmc/collisions.hpp
@@ -56,7 +56,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   int *collision_count);
+                   double *collision_count);
 
 void collide_atoms(int num_atoms,
                    int num_cells,
@@ -71,7 +71,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   int *collision_count);
+                   double *collision_count);
 
 void index_atoms(int num_atoms,
                  double3 *pos,
@@ -107,7 +107,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              curandState *state,
-             int *collision_count,
+             double *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel);
@@ -117,7 +117,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              pcg32_random_t *state,
-             int *collision_count,
+             double *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel);
