+++ src/collisions.cpp
@@ -111,7 +111,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   int *collision_count) {
+                   double *collision_count) {
     // Index atoms
     index_atoms(num_atoms,
                 pos,
@@ -160,7 +160,7 @@ void collide_atoms(int num_atoms,
                    int *cell_num_atoms,
                    int *cell_cumulative_num_atoms,
                    double *collision_remainder,
-                   int *collision_count) {
+                   double *collision_count) {
     // Index atoms
     index_atoms(num_atoms,
                 pos,
@@ -410,7 +410,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              curandState *state,
-             int *collision_count,
+             double *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel) {
@@ -432,7 +432,7 @@ void collide(int num_cells,
              int *cell_cumulative_num_atoms,
              double dt,
              pcg32_random_t *state,
-             int *collision_count,
+             double *collision_count,
              double *collision_remainder,
              double  *sig_vr_max,
              double3 *vel) {
